src/backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
databases==0.9.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
mysql-connector-python==8.3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.1
